Parallel concordance

Parallel concordance

para_conc(
  source_text = "The source text corpus",
  target_text = "The target text corpus",
  pattern = "Search pattern for words in the source text",
  case_insensitive = FALSE,
  conc_sample = 25,
  filename = "parallel_conc.txt"
)

Arguments

source_text
character vector of the source-text corpora

target_text
character vector of the target-text corpora

pattern
regular expression search pattern for the source-text node word

case_insensitive
logical; whether the search pattern is case insensitive. Default to FALSE

conc_sample
random sample of the concordance lines

filename
file name of the parallel concordance output

Value

A tibble of parallel concordance with source-text node word and its left and right context, and their target-text translation

Examples

para_conc(sci_en, sci_id, pattern = "should", conc_sample = 20)

#> The output concordance file (called: 'parallel_conc.txt') will be saved in this directory: '/Users/Primahadi/Documents/r-packages/paracorp/docs/reference'

#> The output concordance will ALSO be returned as a tibble data frame in the R console.

#> Detecting the match/pattern...

#> You choose to generate a 20 random-sample of the concordance lines.

#> Creating a 20 random-sample of the concordance lines...

#> Generating the concordance for the match/pattern...

#> Saving the output concordance file (called: 'parallel_conc.txt') in '/Users/Primahadi/Documents/r-packages/paracorp/docs/reference'.

#> # A tibble: 20 x 4
#>    LEFT                   NODE  RIGHT                 TRANSLATION               
#>    <chr>                  <chr> <chr>                 <chr>                     
#>  1 This enzyme should cr… shou… also react to the ho… "Enzim ini tentunya menja…
#>  2 When designating thes… shou… always be borne in m… "Ketika menentukan filum …
#>  3 It                     shou… be admitted that ini… "Perlu diakui bahwa untuk…
#>  4 The minor improvement… shou… be as readily preser… "Perubahan kecil dari gen…
#>  5 To reach that, a cond… shou… be created, either i… "Untuk menggapainya, haru…
#>  6 This was doubly ironi… shou… be denigrated rather… "Hal ini sangat ironis, k…
#>  7 The number of these t… shou… be found all over th… "Jumlah bentuk peralihan …
#>  8 So, in designing a ta… shou… be identified betwee… "Oleh karena itu dalam pe…
#>  9 With thousands of exa… shou… be possible to descr… "Dengan ribuan contoh mut…
#> 10 It                     shou… be realized that not… "Bukan hanya pembakaran k…
#> 11 When airplane has bee… shou… be reliable so the a… "Ketika pesawat telah jad…
#> 12 So that, in arranging… shou… be seriously conside… "Sehingga dalam menyusun …
#> 13 Nucleus also function… shou… be started, executed… "Nukleus juga berfungsi m…
#> 14 The fibre is for stre… shou… be.                   "Serat untuk menguatkan l…
#> 15 Organization in compa… shou… focus on user produc… "Organisasi di perusahaan…
#> 16 Prior to the genom pr… shou… have been gained, Ar… "Sebelum ada proyek genom…
#> 17 It might seem odd tha… shou… lie at the center of… "Mungkin terlihat aneh ba…
#> 18 This very powerful ma… shou… not be confused with… "Baterai utama yang sanga…
#> 19 Products or services … shou… not only know but al… "Penyedia produk atau jas…
#> 20 Nevertheless Posman r… shou… take precautions whe… "Meski begitu, Posman men…

          # we delete the output file to remove warning in R CMD check
          unlink("parallel_conc.txt")